+++ libs/exporter/src/vault-export/index.ts
@@ -1,2 +1,4 @@
 export * from "./services/vault-export.service.abstraction";
 export * from "./services/vault-export.service";
+export * from "./services/org-vault-export.service.abstraction";
+export * from "./services/org-vault-export.service";
