+++ libs/exporter/src/vault-export/services/vault-export.service.abstraction.ts
@@ -2,8 +2,7 @@ export const EXPORT_FORMATS = ["csv", "json", "encrypted_json"] as const;
 export type ExportFormat = (typeof EXPORT_FORMATS)[number];
 
 export abstract class VaultExportServiceAbstraction {
-  getExport: (format?: ExportFormat, organizationId?: string) => Promise<string>;
-  getPasswordProtectedExport: (password: string, organizationId?: string) => Promise<string>;
-  getOrganizationExport: (organizationId: string, format?: ExportFormat) => Promise<string>;
+  getExport: (format?: ExportFormat) => Promise<string>;
+  getPasswordProtectedExport: (password: string) => Promise<string>;
   getFileName: (prefix?: string, extension?: string) => string;
 }
